Break textureatlas out into its own file.. Make it dynamically resize.
